Fixed some warnings in VC
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk/sphinx3@4211 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ src/libs3audio/cont_ad_base.c
@@ -1043,7 +1043,7 @@ cont_ad_t *cont_ad_init (ad_rec_t *a, int32 (*func)(ad_rec_t *, int16 *, int32))
     r->thresh_sil = r->noise_level + r->delta_sil;
     r->thresh_speech = r->noise_level + r->delta_speech;
     r->thresh_update = CONT_AD_THRESH_UPDATE;
-    r->adapt_rate = CONT_AD_ADAPT_RATE;
+    r->adapt_rate = (float32)CONT_AD_ADAPT_RATE;
     
     r->state = CONT_AD_STATE_SIL;
